requirements.txt
agora-token-builder==1.0.0
arrow==1.2.3
asgiref==3.5.2
astroid==2.3.3
beautifulsoup4==4.11.1
binaryornot==0.4.4
botocore==1.27.71
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.3
cookiecutter==2.1.1
dj-database-url==0.5.0
Django==4.0.6
django-browser-reload==1.6.0
django-cors-headers==3.13.0
django-crispy-forms==1.8.1
django-environ==0.9.0
django-filter==2.2.0
django-login-required-middleware==0.4
django-rest-framework==0.1.0
django-tailwind==3.4.0
django-widget-tweaks==1.4.5
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.4
isort==4.3.21
Jinja2==3.1.2
jinja2-time==0.2.0
jmespath==1.0.1
lazy-object-proxy==1.4.3
MarkupSafe==2.1.1
mccabe==0.6.1
nanoid==2.0.0
Pillow==9.2.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pylint==2.4.4
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
python-slugify==7.0.0
pytz==2022.1
PyYAML==6.0
requests==2.28.1
six==1.13.0
soupsieve==2.3.2.post1
sql==2022.4.0
sqlparse==0.4.2
text-unidecode==1.3
urllib3==1.26.12
whitenoise==6.2.0
wrapt==1.11.2
